pyproject.toml
[project]
name = "adme-autoresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chemprop>=2.2.2",
    "datasets>=4.7.0",
    "pytorch-lightning>=2.6.1",
]
